requirements.txt
streamlit
rembg
pillow
beautifulsoup4
requests
watchdog